error al sacar un item del carrito

+++ Vista/Producto.aspx.cs
@@ -169,7 +169,7 @@ namespace Vista
                 car._articulos[posInCarrito].CantAgregar = 0;
                 lbl_Cantidad.Text = car._articulos[posInCarrito].CantAgregar.ToString();
                 lbl_Agregado.Text = "Has quitado todos los " + car._articulos[posInCarrito].Producto.Nombre + " del carrito";
-                lbl_SubTotal.Text = Convert.ToString(car._articulos[posInCarrito].Cant * Convert.ToInt32(lbl_precio.Text));
+                lbl_SubTotal.Text = Convert.ToString(car._articulos[posInCarrito].Cant * Convert.ToDouble(lbl_precio.Text));
                 lblCantidadActual.Text = car._articulos[posInCarrito].Cant.ToString();
             }else
             {
